feat(react-query): convert to TypeScript & update structure
