+++ packages/react-query/package.json
@@ -44,7 +44,7 @@
     "scripts"
   ],
   "bin": {
-    "suspensive-react-query-switch": "bin/suspensive-react-query-switch.js"
+    "suspensive-react-query-switch": "dist/bin/suspensive-react-query-switch.js"
   },
   "scripts": {
     "build": "tsup",
@@ -55,7 +55,7 @@
     "clean": "rimraf ./dist && rimraf ./coverage",
     "dev": "tsup --watch",
     "prepack": "pnpm build",
-    "postinstall": "node -e \"import('./scripts/postinstall.js').catch(e => console.error(e))\""
+    "postinstall": "node -e \"import('./dist/postinstall.js').catch(e => console.error(e))\""
   },
   "dependencies": {
     "@suspensive/react-query-4": "workspace:^0.0.1",
