+++ packages/react-query/src/bin/suspensive-react-query-switch.ts
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 'use strict'
-import '../scripts/switch.js'
+import '../utils/switch'
