+++ packages/react-query/src/utils/loadModule.ts
@@ -1,0 +1,8 @@
+export function loadModule<T>(name: string): T | undefined {
+  try {
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-return, @typescript-eslint/no-unsafe-argument
+    return require(name)
+  } catch (e) {
+    return undefined
+  }
+}
