+++ packages/react-query/tsup.config.ts
@@ -1,4 +1,4 @@
 import { options } from '@suspensive/tsup'
 import { defineConfig } from 'tsup'
 
-export default defineConfig({ ...options, entry: ['src/**/index.ts'] })
+export default defineConfig({ ...options, entry: ['src/**/*.ts'] })
